+++ src/Exceptions/EntryNotFound.php
@@ -2,9 +2,7 @@
 
 namespace DoubleThreeDigital\SimpleCommerce\Exceptions;
 
-use Exception;
-
-class OrderNotFound extends Exception
+class EntryNotFound extends \Exception
 {
-    //
+
 }